Revert "Extremely Randomized Trees implementation for the SPO loss"
This reverts commit 1005212dc8b5dc01f8455eac34a5ec33edd1f1b5.

+++ Algorithms/SPOForest.py
@@ -89,7 +89,7 @@ class SPOForest(object):
     
     if seed is not None:
       np.random.seed(seed)
-    tree_seeds = np.random.randint(0, high=2**32-1, size=self.n_estimators, dtype=np.int64)
+    tree_seeds = np.random.randint(0, high=2**32-1, size=self.n_estimators)
     
     
     if self.num_workers == 1:
